cs.venti.datadir: DataDir: add new create_statedir and create_datadir optional parameters, support autocreating the default inner "data" subdir
The paramaters create_statedir and create_datadir are Booleans
controlling the autocreation of the statedir or datadir is missings,
respectively. They default to False, save that if the datadirpath
is not specified and the statedirpath exists and the default
datadirpath does not exist, create_datadir defaults to True. In
this way one can create an empty statedir and the implicit "data"
subdirectory will be automatcically be made on first use.

--HG--
branch : venti