cs.vt.block: big refactor of indirect blocks, making them their own class but keeping the F_INDIRECT flag in the superblock's serialisation.
This has a numberof benefits:

Indirect blocks now follow a similar pattern to SubBlocks.

The indirection logic is nicely localised.

All the other Block types become leaf blocks, removing the internal
indirect logic and allowing any of them to serve as the superblock
of an indirect block. This also allows dropping the "indirect"
parameter from their constructors.

One side effect is separation of a Block's .span attribute from the
length of its data; for leaf blocks it matches the length of the
data and for indirect blocks it matches the sum of the subblock
spans.

Assorted other cleanups came with this, such as dropping the .all_data
methods and making .data an on-demand method, usually computed n
__getattr__ from the private ._data() method.

HashCodeBlocks may be constructed with a .span of None, in which
case their spans are computed by fetching the leaf data.

TODO: the text transcription has changed. Still need to add code
to support the legacy IB{...} transcription format.

--HG--
branch : venti