cs.vt.blockmap: big refactor of BlockMap and MappedFD
MappedFD submaps not only know about their hash class, and have
methods to return an offset or a MapEntry namedtuple from a map
entry index.

A MapEntry is a namedtuple with index, offset, span and hashcode
attributes and .leaf and .data convenience properties.

The BlockMap now supports persistent BlockMaps, which have well
named persistent files containing the submap data, and better support
the progressive update where the BlockMap is usable before all the
submaps are complete.

The _MappedFDStub used to terminate the linked list of submaps is
gone, and the submaps are now only a list.

--HG--
branch : venti