cs.vt.file: File: refactor .read to wrap expose the core reading scan as the new .datafrom generator method, and to collect data from a reusable CornuCopyBuffer using datafrom
The .datafrom method is a genarator returning natural slices from
the File's underlying ._file.

The .read method now obtains data from a CornuCopyBuffer using
.datafrom as its chunk source, and keeps a reference to the last
CornuCopyBuffer used. If a subsequent .read resumes at the same
offset as the buffer, the buffer is reused obviating all the "find
the start of the data in the file" logic that would otherwise be
needed since the buffer is reading from a generator which has already
performed that operation. Those costs, particularly for files backed
by indirect Blocks, are substantial and a crude test on such a file
with small leaf Blocks gave a roughly 3 to 4 times speedup in
throughput.

The .readall method is now a trivial stub calling .read with a size of -1.

--HG--
branch : venti