Remove pyicontract-lint
It has been abandoned and is no longer maintained.
It raises this message during install:
DEPRECATION: Building 'pyicontract-lint' using the legacy setup.py bdist_wheel mechanism,
which will be removed in a future version. pip 25.3 will enforce this behaviour change.
A possible replacement is to use the standardized build interface by setting the --use-pep517 option,
(possibly combined with --no-build-isolation),
or adding a pyproject.toml file to the source tree of 'pyicontract-lint'.
Discussion can be found at https://github.com/pypa/pip/issues/6334.

Looks like it was only used in a single line which was commented out,
so I've just removed it.

Signed-off-by: Bradley Reynolds <bradley.reynolds@tailstory.dev>